requirements.txt
python-dotenv
flask
requests
streamlit
beautifulsoup4
langchain-groq
langchain_huggingface
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
langchain-core
ollama
duckduckgo_search
openai